Testing: move `http_test.go` and update to avoid circular import
